DPY PolyBot/requirements.txt
discord.py
aiohttp
cairosvg
chess
matplotlib
motor
yt-dlp
pynacl
discord-components